moving some functions out of main.cpp into util.cpp
